+++ sdk/core/azure-core/inc/azure/core/context.hpp
@@ -10,11 +10,11 @@
 
 #include "azure/core/azure_assert.hpp"
 #include "azure/core/datetime.hpp"
+#include "azure/core/dll_import_export.hpp"
 
 #include <atomic>
 #include <chrono>
 #include <memory>
-#include <new> //For the non-allocating placement new
 #include <stdexcept>
 #include <string>
 #include <type_traits>
@@ -198,7 +198,7 @@ namespace Azure { namespace Core {
      * @brief Check if the context is cancelled.
      * @return `true` if this context is cancelled, `false` otherwise.
      */
-    bool IsCancelled() const { return GetDeadline() < std::chrono::system_clock::now(); }
+    bool IsCancelled() const noexcept { return GetDeadline() < std::chrono::system_clock::now(); }
 
     /**
      * @brief Throw an exception if the context was cancelled.
@@ -212,8 +212,8 @@ namespace Azure { namespace Core {
     }
 
     /**
-     * @brief Get the application context (root).
+     * @brief The application context (root).
      */
-    static Context& GetApplicationContext();
+    static AZ_CORE_DLLEXPORT Context ApplicationContext;
   };
 }} // namespace Azure::Core